fix(pipeline/odpt): revert agency_id to provider.name.en.long for route consistency
Commit 28474d8 changed buildAgencyV2 to use `provider.name.en.short`
in the agency_id formula, but buildRoutesV2 still uses
`provider.name.en.long` for `RouteV2Json.ai`. When long != short,
routes would reference an agency_id that no agency record carries,
breaking route -> agency lookup in mergeSourcesV2.

Latent in current production: only Yurikamome is an ODPT train
source and its provider config has `long === short === 'Yurikamome'`,
so both formulas produce the same `yurimo:Yurikamome` by accident.
Verified the existing `public/data-v2/yurimo/data.json` matches the
reverted formula (no data rebuild required to stay consistent).

Revert buildAgencyV2 to `provider.name.en.long` so it matches
buildRoutesV2. Add a TSDoc note cross-referencing buildRoutesV2 so
a future edit to either side is forced to consider the other. Update
the test fixture expected values accordingly.

Refs: #121 (Copilot review comment on build-agency.ts)

+++ pipeline/src/lib/pipeline/app-data-v2/odpt/build-agency.ts
@@ -13,6 +13,10 @@ import type { Provider } from '../../../../types/resource-common';
 /**
  * Build AgencyV2Json[] from provider info.
  *
+ * The agency_id scheme (`${prefix}:${provider.name.en.long}`) must stay
+ * in sync with {@link buildRoutesV2}, which sets `RouteV2Json.ai` using
+ * the same formula so that route → agency lookups succeed after merge.
+ *
  * @param prefix - Source prefix for ID namespacing.
  * @param provider - Provider info used only for agency_id generation.
  * @returns Array with a single AgencyV2Json record.
@@ -21,7 +25,7 @@ export function buildAgencyV2(prefix: string, provider: Provider): AgencyV2Json[
   return [
     {
       v: 2,
-      i: `${prefix}:${provider.name.en.short}`,
+      i: `${prefix}:${provider.name.en.long}`,
       n: '',
       u: '',
       tz: 'Asia/Tokyo',
